test(cmd): move expand-helper tests to expand_test.go
Follow-up to moving the expand helper functions from create.go to
expand.go: relocate their unit tests (setCLIVariables, setBackendConfig,
setValidationLevel, validateMaybeDie) from create_test.go into a new
expand_test.go so the tests live next to the code under test. Pure
relocation, no test logic changes. Addresses review feedback on #5781.
